data_generator/pyproject.toml
[project]
name = "data-generator"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "faker>=40.5.1",
    "faker-commerce>=1.0.4",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.47",
]
